push pop and peek complete
