+++ coding-challenges/401/stacksAndQueues/node.js
@@ -1,0 +1,10 @@
+'use strict';
+
+class Node{
+  constructor (value){
+    this.value = value;
+  }
+}
+
+module.exports = Node;
+
