--- coding-challenges/401/stacksAndQueues/queue.js
@@ -1,19 +1,0 @@
-class Queue{
-  constructor(){
-    this.q = new Array();
-    this.front = null;
-  }
-
-  peek(){
-    return this.front;
-  }
-
-  enqueue(value) {
-    //puts in the rear of the queue
-  }
-
-  dequeue() {
-    //removes the front of the queue
-  }
-
-}
